Ci/dev (#117)
* moving signature


* Added operator_info


* Orange test is doing OPP_FORCE_INDEX_UPDATE=1


* Cluster_type


* OPP_AUTO_PACKAGEMANIFEST_CLUSTER_VERSION_LABEL is set


* Test


* Test


* Test


* Operator info


* Fix


* Fix


* Fix


* Fix


* Fix


* Fix


* Test


* Fix


* Comment added


* Comment added


* Fix


* Improved comment


* Fix


* Test


* Fix


* Fixed push to quay


* Test


* Test


* Imporved strict


* Test


* Fix


* Test

+++ ci/scripts/opp-update.sh
@@ -16,14 +16,16 @@ OPP_ANSIBLE_EXTRA_ARGS=""
 
 OPP_INDEX_IMAGE_POSTFIX=${OPP_INDEX_IMAGE_POSTFIX-"s"}
 
-[ "$1" = "reset" ] && OPP_ANSIBLE_EXTRA_ARGS="-e empty_index=quay.io/operator_testing/index_empty"
+[ "$1" = "reset" ] && OPP_ANSIBLE_EXTRA_ARGS="-e empty_index=quay.io/operator_testing/index_empty" && shift
 
 OPP_TMP_DIR="/tmp/opp-update"
 
 if [ -n "$1" ];then
     cd ci
-    [ -f pipeline-config-${CLUSTER_TYPE}.yaml ] || mv pipeline-config.yaml ci/pipeline-config-${CLUSTER_TYPE}.yaml
-    ln -sfn pipeline-config-${CLUSTER_TYPE}.yaml pipeline-config.yaml
+    if [ ! -f pipeline-config-${CLUSTER_TYPE}.yaml ]; then
+        mv pipeline-config.yaml ci/pipeline-config-${CLUSTER_TYPE}.yaml
+        ln -sfn pipeline-config-${CLUSTER_TYPE}.yaml pipeline-config.yaml
+    fi
     cd -
     CLUSTER_TYPE="-$CLUSTER_TYPE"
 fi
